+++ src/row.c
@@ -267,7 +267,7 @@ int print_row_separator(char *buffer, size_t buffer_sz,
     }
 
 
-    int written = 0;
+    size_t written = 0;
     int tmp = 0;
 
     enum ft_row_type lower_row_type = FT_ROW_COMMON;
@@ -404,7 +404,7 @@ int print_row_separator(char *buffer, size_t buffer_sz,
 
     CHCK_RSLT_ADD_TO_WRITTEN(snprint_n_strings_(buffer + written, buffer_sz - written, 1, "\n"));
 
-    status = written;
+    status = (int)written;
 
 clear:
     F_FREE(top_row_types);
@@ -454,7 +454,7 @@ int wprint_row_separator(wchar_t *buffer, size_t buffer_sz,
     }
 
 
-    int written = 0;
+    size_t written = 0;
     int tmp = 0;
 
     enum ft_row_type lower_row_type = FT_ROW_COMMON;
@@ -591,7 +591,7 @@ int wprint_row_separator(wchar_t *buffer, size_t buffer_sz,
 
     CHCK_RSLT_ADD_TO_WRITTEN(snprint_n_strings_(buffer + written, buffer_sz - written, 1, "\n"));
 
-    status = written;
+    status = (int)written;
 
 clear:
     F_FREE(top_row_types);
@@ -982,7 +982,7 @@ int snprintf_row(const fort_row_t *row, char *buffer, size_t buf_sz, size_t *col
     const char **R = &(*bord_chars)[RR_bip];
 
 
-    int written = 0;
+    size_t written = 0;
     int tmp = 0;
     size_t i = 0;
     for (i = 0; i < row_height; ++i) {
@@ -1029,7 +1029,7 @@ int snprintf_row(const fort_row_t *row, char *buffer, size_t buf_sz, size_t *col
         /* Print new line character */
         CHCK_RSLT_ADD_TO_WRITTEN(snprint_n_strings_(buffer + written, buf_sz - written, 1, new_line_char));
     }
-    return written;
+    return (int)written;
 
 clear:
     return -1;
@@ -1070,7 +1070,7 @@ int wsnprintf_row(const fort_row_t *row, wchar_t *buffer, size_t buf_sz, size_t 
     const char **R = &(*bord_chars)[RR_bip];
 
 
-    int written = 0;
+    size_t written = 0;
     int tmp = 0;
     size_t i = 0;
     for (i = 0; i < row_height; ++i) {
@@ -1117,7 +1117,7 @@ int wsnprintf_row(const fort_row_t *row, wchar_t *buffer, size_t buf_sz, size_t 
         /* Print new line character */
         CHCK_RSLT_ADD_TO_WRITTEN(snprint_n_strings_(buffer + written, buf_sz - written, 1, new_line_char));
     }
-    return written;
+    return (int)written;
 
 clear:
     return -1;
